Added Blog, Entry and Tag entities
Generated Blog, Entry and Tag entities from a jdl

+++ src/MyEnterpriseBlazor.Infrastructure/Data/Repositories/GenericRepository.cs
@@ -110,7 +110,7 @@ namespace MyEnterpriseBlazor.Infrastructure.Data.Repositories
             return fluentRepository;
         }
 
-        protected async Task RemoveManyToManyRelationship(string joinEntityName, string ownerIdKey, string ownedIdKey, long ownerEntityId, List<long> idsToIgnore)
+        protected async Task RemoveManyToManyRelationship(string joinEntityName, string ownerIdKey, string ownedIdKey, int ownerEntityId, List<int> idsToIgnore)
         {
             DbSet<Dictionary<string, object>> dbset = _context.Set<Dictionary<string, object>>(joinEntityName);
